use a different database name for each public key.
this fixes the issue with people (me) using different seeds with the same datadir,
which would cause them to use the same database.

now for each seed/mnemonic a different database will be created at $datadir/db-{pubkeyfirst6chars}.sqlite.

this requires a change on immortan.

considering the fact that everything will be broken anyway we will use this same next release to change other things in the immortan database schema.